feat: Refactor Header component to utilize custom hooks for import/export and header actions. Improved state management and reduced direct atom usage for better code organization and readability.
